some store refactoring
